swagger 적용

+++ src/main/java/com/iam/shopping/service/MemberService.java
@@ -21,12 +21,19 @@ public class MemberService {
         return memberMapper.selectMember();
     }
 
-    public void memberRegister(Member member) {
-        memberMapper.insertMember(member);
+    public void memberRegister(MemberDTO memberDTO) {
+        Map<String, Object> memberMap = Map.of("id", memberDTO.getId()
+                , "pass", memberDTO.getPass()
+                , "name", memberDTO.getName()
+                , "phone", memberDTO.getPhone());
+        if (memberDTO.isTest()) return;
+        int isSuccess = memberMapper.insertMember(memberMap);
     }
 
     public void memberDelete(MemberDTO memberDTO) {
         Map<String, Object> memberMap = Map.of("id", memberDTO.getId());
-        memberMapper.deleteMember(memberMap);
+        System.out.println("삭제여부>" + memberDTO.isTest());
+        if (memberDTO.isTest()) return;
+        int isSuccess = memberMapper.deleteMember(memberMap);
     }
 }
